auth/storage: support for the Accounts Authorizer
In an ideal world this wouldn't be necessary, but I digress.

Unfortunately it appears that the Accounts Data Plane API uses a different method to calculate
the shared key authorization verses other (blob etc) operations which support this auth method.

This commit introduces support for a "from account" type which includes support for this auth
token, so that this authorizer can be used for those endpoints.